codefunctions, but agent dosen't seem to learn much

+++ exercice_4_cartpole_naive_deep_Q/q_network.py
@@ -23,7 +23,7 @@ class Qnetwork(nn.Module):
 
 	# forward propagation
     def forward(self, state):
-        layer1 = F.relu(self.fc1(state))
+        layer1 = F.relu(self.fc1(state.float()))
         layer2 = self.fc2(layer1)
 
         return layer2
